Changed numpy operations on scalars to Python builtin/math operations - np.floor -> math.floor - np.ceil -> math.ceil - np.abs -> abs (could also be done for array operations with no performance hit) - np.round -> round - np.max -> max
Performance improvements range from 10x - 50x

+++ pypulseq/calc_rf_center.py
@@ -26,6 +26,6 @@ def calc_rf_center(rf: SimpleNamespace) -> Tuple[float, float]:
     rf_max = np.max(np.abs(rf.signal))
     i_peak = np.where(np.abs(rf.signal) >= rf_max * 0.99999)[0]
     time_center = (rf.t[i_peak[0]] + rf.t[i_peak[-1]]) / 2
-    id_center = i_peak[int(np.round((len(i_peak) - 1) / 2))]
+    id_center = i_peak[round((len(i_peak) - 1) / 2)]
 
     return time_center, id_center
